Move imagestream engine also to engines directory
